loosened up the term uncle to accomodate farther away blocks
now a nephew is also considered an uncle, given that the parent of a nephew (the uncle) is in the same set as the
nephew. children of nephews will also be uncles.

commented out checks that raise UncleTooOld and UncleIsBrother errors, as those are based on block number of the cannon
chain and would not apply correctly here